chore: move tools into agent builder file to hinder cyclic imports
