+++ zammad-ai/app/answer/__init__.py
@@ -2,8 +2,7 @@
 Answer module for Zammad AI, responsible for generating structured responses to user questions using an agent-based approach.
 """
 
-from .agent import StructuredAgentResponse
-from .context import AgentContext
+from .agent import AgentContext, StructuredAgentResponse
 from .service import AnswerService, get_answer_service
 
 __all__: list[str] = [
