--- zammad-ai/app/answer/context.py
@@ -1,10 +1,0 @@
-from pydantic import BaseModel
-
-from app.qdrant import QdrantKBClient
-
-from .dlf import DLFClient
-
-
-class AgentContext(BaseModel):
-    qdrant_kb_client: QdrantKBClient
-    dlf_client: DLFClient
